# docs/swagger.yaml
basePath: /
definitions:
  dto.AvailabilityData:
    properties:
      available_slots:
        items:
          $ref: '#/definitions/dto.TimeSlot'
        type: array
      booked_slots:
        items:
          $ref: '#/definitions/dto.BookedSlot'
        type: array
      date:
        type: string
      studio_id:
        type: integer
    type: object
  dto.AvailabilityResponse:
    properties:
      available:
        type: boolean
      data:
        $ref: '#/definitions/dto.AvailabilityData'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.BookedSlot:
    properties:
      booking_id:
        type: integer
      end_time:
        type: string
      start_time:
        type: string
    type: object
...
  dto.CheckAvailabilityRequest:
    properties:
      date:
        description: 'Format: "2025-11-20"'
        type: string
      end_time:
        description: 'Format: "17:00"'
        type: string
      start_time:
        description: 'Format: "14:00"'
        type: string
    required:
    - date
    - end_time
    - start_time
    type: object
...
  dto.CreateStudioRequest:
    properties:
      description:
        type: string
      facilities:
        items:
          type: string
        type: array
      image_url:
        type: string
      location:
        type: string
      name:
        minLength: 3
        type: string
      operating_hours:
        description: 'Format: "09:00-22:00"'
        type: string
      price_per_hour:
        minimum: 10000
        type: integer
    required:
    - description
    - facilities
    - image_url
    - location
    - name
    - operating_hours
    - price_per_hour
    type: object
  dto.CreateStudioResponse:
    properties:
      data:
        $ref: '#/definitions/dto.StudioData'
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.DeleteStudioResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
...
  dto.Pagination:
    properties:
      current_page:
        type: integer
      page_size:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
...
  dto.PatchStudioRequest:
    properties:
      description:
        type: string
      facilities:
        items:
          type: string
        type: array
      image_url:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
      operating_hours:
        type: string
      price_per_hour:
        type: integer
    type: object
  dto.PatchStudioResponse:
    properties:
      data:
        $ref: '#/definitions/dto.StudioData'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.StudioData:
    properties:
      created_at:
        type: string
      description:
        type: string
      facilities:
        items:
          type: string
        type: array
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        type: string
      operating_hours:
        type: string
      price_per_hour:
        type: integer
      updated_at:
        type: string
    type: object
  dto.StudioListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.StudioData'
        type: array
      pagination:
        $ref: '#/definitions/dto.Pagination'
      success:
        type: boolean
    type: object
  dto.StudioResponse:
    properties:
      data:
        $ref: '#/definitions/dto.StudioData'
      success:
        type: boolean
    type: object
  dto.TimeSlot:
    properties:
      end_time:
        description: '"12:00"'
        type: string
      start_time:
        description: '"09:00"'
        type: string
    type: object
...
  dto.UpdateStudioRequest:
    properties:
      description:
        type: string
      facilities:
        items:
          type: string
        type: array
      image_url:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      name:
        minLength: 3
        type: string
      operating_hours:
        type: string
      price_per_hour:
        minimum: 10000
        type: integer
    type: object
  dto.UpdateStudioResponse:
    properties:
      data:
        $ref: '#/definitions/dto.StudioData'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Dokumentasi API Backend Booking Studio
  title: Backend Booking Studio API
  version: "1.0"
paths:
...
  /studios:
    get:
      consumes:
      - application/json
      description: Mengambil daftar semua studio dengan filter dan pagination
      parameters:
      - description: Filter lokasi
        in: query
        name: location
        type: string
      - description: Harga minimal
        in: query
        name: min_price
        type: integer
      - description: Harga maksimal
        in: query
        name: max_price
        type: integer
      - description: Hanya studio aktif
        in: query
        name: is_active
        type: boolean
      - description: Cari berdasarkan nama
        in: query
        name: search
        type: string
      - default: 1
        description: Halaman
        in: query
        name: page
        type: integer
      - default: 10
        description: Jumlah data per halaman
        in: query
        name: limit
        type: integer
      - description: Sortir (price_asc, price_desc, name_asc, name_desc)
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StudioListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Ambil semua studio
      tags:
      - Studios
    post:
      consumes:
      - application/json
      description: Menambahkan studio baru oleh admin
      parameters:
      - description: Data studio baru
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStudioRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateStudioResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden (bukan admin)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Buat studio baru (Admin Only)
      tags:
      - Studios
  /studios/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus studio berdasarkan ID
      parameters:
      - description: ID Studio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteStudioResponse'
        "400":
          description: Invalid studio ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden (bukan admin)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Studio not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Hapus studio (Admin Only)
      tags:
      - Studios
    get:
      consumes:
      - application/json
      description: Mengambil detail studio berdasarkan ID
      parameters:
      - description: ID Studio
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StudioResponse'
        "400":
          description: Invalid studio ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Studio not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Ambil 1 studio berdasarkan ID
      tags:
      - Studios
    patch:
      consumes:
      - application/json
      description: Mengupdate sebagian field studio
      parameters:
      - description: ID Studio
        in: path
        name: id
        required: true
        type: integer
      - description: Data patch studio
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.PatchStudioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PatchStudioResponse'
        "400":
          description: Invalid studio ID / payload
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden (bukan admin)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Studio not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Patch sebagian data studio (Admin Only)
      tags:
      - Studios
    put:
      consumes:
      - application/json
      description: Mengupdate seluruh data studio
      parameters:
      - description: ID Studio
        in: path
        name: id
        required: true
        type: integer
      - description: Data update studio
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStudioRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateStudioResponse'
        "400":
          description: Invalid studio ID / payload
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden (bukan admin)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Studio not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update studio lengkap (Admin Only)
      tags:
      - Studios
  /studios/{id}/availability:
    post:
      consumes:
      - application/json
      description: Mengecek apakah studio tersedia pada waktu tertentu
      parameters:
      - description: ID Studio
        in: path
        name: id
        required: true
        type: integer
      - description: Data untuk pengecekan jadwal
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CheckAvailabilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvailabilityResponse'
        "400":
          description: Invalid request payload / invalid studio ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Cek jadwal ketersediaan studio
      tags:
      - Studios
swagger: "2.0"